Fixed minor issues, testing done, ready for merge
