+++ web/src/routes/_auth/music/index.tsx
@@ -26,6 +26,7 @@ import { Spinner } from "@/components/ui/spinner";
 import { useContentFadeTransition } from "@/hooks/useContentFadeTransition";
 import { useWindowVirtualizer } from "@tanstack/react-virtual";
 import { useVirtualizedInfiniteLoader } from "@/hooks/useVirtualizedInfiniteLoader";
+import { useWindowScrollMargin } from "@/hooks/useWindowScrollMargin";
 import { showActionFailed } from "@/lib/toast-helpers";
 import LiveAnnouncer from "@/components/LiveAnnouncer";
 import { unwrapString, unwrapInt, unwrapStringOrUndefined } from "@/lib/nullable";
@@ -653,40 +654,7 @@ function VirtualizedTracksList({
 }: VirtualizedTracksListProps) {
   "use no memo";
 
-  const listRef = useRef<HTMLDivElement>(null);
-  const [scrollMargin, setScrollMargin] = useState(0);
-
-  // The app shell scrolls the window (not an inner container), so virtualize
-  // against the window. scrollMargin is the list's distance from the top of the
-  // document, letting the virtualizer map window scroll onto item indices.
-  useEffect(() => {
-    const listElement = listRef.current;
-    if (!listElement) {
-      return;
-    }
-
-    const updateScrollMargin = () => {
-      setScrollMargin(listElement.getBoundingClientRect().top + window.scrollY);
-    };
-
-    updateScrollMargin();
-
-    const resizeObserver =
-      typeof ResizeObserver === "undefined"
-        ? null
-        : new ResizeObserver(updateScrollMargin);
-
-    resizeObserver?.observe(listElement);
-    if (typeof document !== "undefined") {
-      resizeObserver?.observe(document.body);
-    }
-    window.addEventListener("resize", updateScrollMargin);
-
-    return () => {
-      resizeObserver?.disconnect();
-      window.removeEventListener("resize", updateScrollMargin);
-    };
-  }, []);
+  const { listRef, scrollMargin } = useWindowScrollMargin<HTMLDivElement>();
 
   const onChange = useVirtualizedInfiniteLoader({
     itemCount: virtualItems.length,
